perl run_abyss.pl for multi k-mer assemblies, merging assemblies, and generated assembly stats
This script runs an installation of ABySS-1.3.4 on beocat. Step 1 is
the assembly of multiple single k-mer assemblies. Step 2 is the merging
of these single k-mer assemblies by running ABySS on reads and the
unitigs from the multi-kmer assemblies. Step 3 writes a script to
calculate N25,N50,N75, cumulative scaffold length, number of scaffolds
for all assemblies.